raspberry_pi_3b/01_trash_notifier_e_ink/requirements.txt
arrow==1.3.0
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
colorzero==2.0
gpiozero==2.0.1
ics==0.7.2
idna==3.10
lgpio==0.2.2.0
pillow==11.0.0
python-dateutil==2.9.0.post0
requests==2.32.3
RPi.GPIO==0.7.1
six==1.16.0
spi==0.2.0
spidev==3.6
TatSu==5.12.2
types-python-dateutil==2.9.0.20241003
urllib3==2.2.3
